# swagger.yaml
openapi: 3.0.0
info:
  title: Email Notification System
  description: API documentation for sending emails
  version: 1.0.0
paths:
  /api/emails/send:
    post:
      summary: Send an email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                receiverName:
                  type: string
                  example: reciepientName
                receiverAddress:
                  type: string
                  example: recipient@example.com
                subject:
                  type: string
                  example: Test Email
                textMessage:
                  type: string
                  example: Hello, this is a test email!
      responses:
        "200":
          description: Email queued for sending
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 66914fce018a6534c894ce39

                  message:
                    type: string
                    example: Email queued for sending
        "400":
          description: Invalid request body
        "500":
          description: Internal server error

  /api/emails/status/{emailId}:
    get:
      summary: Get email status by emailId
      parameters:
        - name: emailId
          in: path
          required: true
          schema:
            type: string
            example: 66914fce018a6534c894ce39
          description: Email ID to retrieve status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 66914fce018a6534c894ce39

                  status:
                    type: string
                    example: sent
        "404":
          description: Email not found
        "500":
          description: Internal server error